fixed compiler's warnings. fixed bug with nsalert which was launched in not main thread.

+++ NSArray+Color.m
@@ -26,7 +26,7 @@
 
 +(NSArray*)		arrayWithColor: (NSColor*) col
 {
-	float			fRed, fGreen, fBlue, fAlpha;
+	CGFloat	fRed, fGreen, fBlue, fAlpha;
 	
 	col = [col colorUsingColorSpaceName: NSCalibratedRGBColorSpace];
 	[col getRed: &fRed green: &fGreen blue: &fBlue alpha: &fAlpha];
